Remove unnecessary candidate system spec
This spec tests redirect behaviour on a new application form, merge it
with the spec that deals with submitting an application.
